Struct bdk_wallet::descriptor::Miniscript

pub struct Miniscript<Pk, Ctx>
where
    Pk: MiniscriptKey,
    Ctx: ScriptContext,
{
    pub node: Terminal<Pk, Ctx>,
    pub ty: Type,
    pub ext: ExtData,
    /* private fields */
}

The top-level miniscript abstract syntax tree (AST).

Fields
node: Terminal<Pk, Ctx>

A node in the AST.

ty: Type

The correctness and malleability type information for the AST node.

ext: ExtData

Additional information helpful for extra analysis.

Implementations

impl<Pk, Ctx> Miniscript<Pk, Ctx>
where Pk: MiniscriptKey, Ctx: ScriptContext,

pub fn requires_sig(&self) -> bool

Whether all spend paths of miniscript require a signature

pub fn is_non_malleable(&self) -> bool

Whether the miniscript is malleable

pub fn within_resource_limits(&self) -> bool

Whether the miniscript can exceed the resource limits(Opcodes, Stack limit etc)

pub fn has_mixed_timelocks(&self) -> bool

Whether the miniscript contains a combination of timelocks

pub fn has_repeated_keys(&self) -> bool

Whether the miniscript has repeated Pk or Pkh

pub fn contains_raw_pkh(&self) -> bool

Whether the given miniscript contains a raw pkh fragment

pub fn sanity_check(&self) -> Result<(), AnalysisError>

Check whether the underlying Miniscript is safe under the current context Lifting these polices would create a semantic representation that does not represent the underlying semantics when miniscript is spent. Signing logic may not find satisfaction even if one exists.

For most cases, users should be dealing with safe scripts. Use this function to check whether the guarantees of library hold. Most functions of the library like would still work, but results cannot be relied upon

pub fn ext_check(&self, ext: &ExtParams) -> Result<(), AnalysisError>

Check whether the miniscript follows the given Extra policy [ExtParams]

impl<Pk, Ctx> Miniscript<Pk, Ctx>
where Pk: MiniscriptKey, Ctx: ScriptContext,

Iterator-related extensions for Miniscript

pub fn iter(&self) -> Iter<'_, Pk, Ctx>

Creates a new [Iter] iterator that will iterate over all Miniscript items within AST by traversing its branches. For the specific algorithm please see [Iter::next] function.

pub fn iter_pk(&self) -> PkIter<'_, Pk, Ctx>

Creates a new [PkIter] iterator that will iterate over all plain public keys (and not key hash values) present in Miniscript items within AST by traversing all its branches. For the specific algorithm please see [PkIter::next] function.

pub fn branches(&self) -> Vec<&Miniscript<Pk, Ctx>>

Enumerates all child nodes of the current AST node (self) and returns a Vec referencing them.

pub fn get_nth_child(&self, n: usize) -> Option<&Miniscript<Pk, Ctx>>

Returns child node with given index, if any

pub fn get_nth_pk(&self, n: usize) -> Option<Pk>

Returns Option::Some with cloned n’th public key from the current miniscript item, if any. Otherwise returns Option::None.

NB: The function analyzes only single miniscript item and not any of its descendants in AST.

impl<Pk, Ctx> Miniscript<Pk, Ctx>
where Pk: MiniscriptKey, Ctx: ScriptContext,

pub const TRUE: Miniscript<Pk, Ctx> = _

The 1 combinator.

pub const FALSE: Miniscript<Pk, Ctx> = _

The 0 combinator.

pub fn from_ast(t: Terminal<Pk, Ctx>) -> Result<Miniscript<Pk, Ctx>, Error>

Add type information(Type and Extdata) to Miniscript based on AstElem fragment. Dependent on display and clone because of Error Display code of type_check.

pub fn from_components_unchecked( node: Terminal<Pk, Ctx>, ty: Type, ext: ExtData ) -> Miniscript<Pk, Ctx>

Create a new Miniscript from a Terminal node and a Type annotation This does not check the typing rules. The user is responsible for ensuring that the type provided is correct.

You should almost always use Miniscript::from_ast instead of this function.

impl<Pk, Ctx> Miniscript<Pk, Ctx>
where Pk: MiniscriptKey, Ctx: ScriptContext,

pub fn into_inner(self) -> Terminal<Pk, Ctx>

Extracts the AstElem representing the root of the miniscript

pub fn as_inner(&self) -> &Terminal<Pk, Ctx>

Get a reference to the inner AstElem representing the root of miniscript

pub fn encode(&self) -> ScriptBuf
where Pk: ToPublicKey,

Encode as a Bitcoin script

pub fn script_size(&self) -> usize

Size, in bytes of the script-pubkey. If this Miniscript is used outside of segwit (e.g. in a bare or P2SH descriptor), this quantity should be multiplied by 4 to compute the weight.

In general, it is not recommended to use this function directly, but to instead call the corresponding function on a Descriptor, which will handle the segwit/non-segwit technicalities for you.

pub fn max_satisfaction_witness_elements(&self) -> Result<usize, Error>

Maximum number of witness elements used to satisfy the Miniscript fragment, including the witness script itself. Used to estimate the weight of the VarInt that specifies this number in a serialized transaction.

This function may returns Error when the Miniscript is impossible to satisfy

pub fn max_satisfaction_size(&self) -> Result<usize, Error>

Maximum size, in bytes, of a satisfying witness. For Segwit outputs one_cost should be set to 2, since the number 1 requires two bytes to encode. For non-segwit outputs one_cost should be set to 1, since OP_1 is available in scriptSigs.

In general, it is not recommended to use this function directly, but to instead call the corresponding function on a Descriptor, which will handle the segwit/non-segwit technicalities for you.

All signatures are assumed to be 73 bytes in size, including the length prefix (segwit) or push opcode (pre-segwit) and sighash postfix.

pub fn satisfy<S>(&self, satisfier: S) -> Result<Vec<Vec<u8>>, Error>
where S: Satisfier<Pk>, Pk: ToPublicKey,

Attempt to produce non-malleable satisfying witness for the witness script represented by the parse tree

pub fn satisfy_malleable<S>(&self, satisfier: S) -> Result<Vec<Vec<u8>>, Error>
where S: Satisfier<Pk>, Pk: ToPublicKey,

Attempt to produce a malleable satisfying witness for the witness script represented by the parse tree

pub fn build_template<P>(&self, provider: &P) -> Satisfaction<Placeholder<Pk>>
where P: AssetProvider<Pk>, Pk: ToPublicKey,

Attempt to produce a non-malleable witness template given the assets available

pub fn build_template_mall<P>( &self, provider: &P ) -> Satisfaction<Placeholder<Pk>>
where P: AssetProvider<Pk>, Pk: ToPublicKey,

Attempt to produce a malleable witness template given the assets available

impl<Ctx> Miniscript<<Ctx as ScriptContext>::Key, Ctx>
where Ctx: ScriptContext,

pub fn parse_insane( script: &Script ) -> Result<Miniscript<<Ctx as ScriptContext>::Key, Ctx>, Error>

Attempt to parse an insane(scripts don’t clear sanity checks) script into a Miniscript representation. Use this to parse scripts with repeated pubkeys, timelock mixing, malleable scripts without sig or scripts that can exceed resource limits. Some of the analysis guarantees of miniscript are lost when dealing with insane scripts. In general, in a multi-party setting users should only accept sane scripts.

pub fn parse_with_ext( script: &Script, ext: &ExtParams ) -> Result<Miniscript<<Ctx as ScriptContext>::Key, Ctx>, Error>

Attempt to parse an miniscript with extra features that not yet specified in the spec. Users should not use this function unless they scripts can/will change in the future. Currently, this function supports the following features: - Parsing all insane scripts - Parsing miniscripts with raw pubkey hashes

Allowed extra features can be specified by the ext [ExtParams] argument.

pub fn parse( script: &Script ) -> Result<Miniscript<<Ctx as ScriptContext>::Key, Ctx>, Error>

Attempt to parse a Script into Miniscript representation.

This function will fail parsing for scripts that do not clear the Miniscript::sanity_check checks. Use Miniscript::parse_insane to parse such scripts.

Decode/Parse a miniscript from script hex

use miniscript::{Miniscript, Segwitv0, Tap};
use miniscript::bitcoin::secp256k1::XOnlyPublicKey;
use miniscript::bitcoin::hashes::hex::FromHex;

type Segwitv0Script = Miniscript<bitcoin::PublicKey, Segwitv0>;
type TapScript = Miniscript<XOnlyPublicKey, Tap>;

// parse x-only miniscript in Taproot context
let tapscript_ms = TapScript::parse(&bitcoin::ScriptBuf::from_hex(
    "202788ee41e76f4f3af603da5bc8fa22997bc0344bb0f95666ba6aaff0242baa99ac",
).expect("Even length hex"))
    .expect("Xonly keys are valid only in taproot context");
// tapscript fails decoding when we use them with compressed keys
let err = TapScript::parse(&bitcoin::ScriptBuf::from_hex(
    "21022788ee41e76f4f3af603da5bc8fa22997bc0344bb0f95666ba6aaff0242baa99ac",
).expect("Even length hex"))
    .expect_err("Compressed keys cannot be used in Taproot context");
// Segwitv0 succeeds decoding with full keys.
Segwitv0Script::parse(&bitcoin::ScriptBuf::from_hex(
    "21022788ee41e76f4f3af603da5bc8fa22997bc0344bb0f95666ba6aaff0242baa99ac",
).expect("Even length hex"))
    .expect("Compressed keys are allowed in Segwit context");

impl<Pk, Ctx> Miniscript<Pk, Ctx>
where Pk: MiniscriptKey, Ctx: ScriptContext,

pub fn substitute_raw_pkh( &self, pk_map: &BTreeMap<Hash, Pk> ) -> Miniscript<Pk, Ctx>

Substitutes raw public keys hashes with the public keys as provided by map.

impl<Pk, Ctx> Miniscript<Pk, Ctx>
where Pk: FromStrKey, Ctx: ScriptContext,

pub fn from_str_insane(s: &str) -> Result<Miniscript<Pk, Ctx>, Error>

Attempt to parse an insane(scripts don’t clear sanity checks) from string into a Miniscript representation. Use this to parse scripts with repeated pubkeys, timelock mixing, malleable scripts without sig or scripts that can exceed resource limits. Some of the analysis guarantees of miniscript are lost when dealing with insane scripts. In general, in a multi-party setting users should only accept sane scripts.

pub fn from_str_ext( s: &str, ext: &ExtParams ) -> Result<Miniscript<Pk, Ctx>, Error>

Attempt to parse an Miniscripts that don’t follow the spec. Use this to parse scripts with repeated pubkeys, timelock mixing, malleable scripts, raw pubkey hashes without sig or scripts that can exceed resource limits.

Use [ExtParams] builder to specify the types of non-sane rules to allow while parsing.

impl<Pk, Ctx> Miniscript<Pk, Ctx>
where Pk: MiniscriptKey, Ctx: ScriptContext,

pub fn lift_check(&self) -> Result<(), LiftError>

Lifting corresponds to conversion of a miniscript into a [Semantic] policy for human readable or machine analysis. However, naively lifting miniscripts can result in incorrect interpretations that don’t correspond to the underlying semantics when we try to spend them on bitcoin network. This can occur if the miniscript contains:

A combination of timelocks

A spend that exceeds resource limits