bdk_wallet::descriptor
Module template

Descriptor templates

This module contains the definition of various common script templates that are ready to be used. See the documentation of each template for an example.

Structs

Bip44

BIP44 template. Expands to pkh(key/44'/{0,1}'/0'/{0,1}/*)

Bip49

BIP49 template. Expands to sh(wpkh(key/49'/{0,1}'/0'/{0,1}/*))

Bip84

BIP84 template. Expands to wpkh(key/84'/{0,1}'/0'/{0,1}/*)

Bip86

BIP86 template. Expands to tr(key/86'/{0,1}'/0'/{0,1}/*)

Bip44Public

BIP44 public template. Expands to pkh(key/{0,1}/*)

Bip49Public

BIP49 public template. Expands to sh(wpkh(key/{0,1}/*))

Bip84Public

BIP84 public template. Expands to wpkh(key/{0,1}/*)

Bip86Public

BIP86 public template. Expands to tr(key/{0,1}/*)

P2Pkh

P2PKH template. Expands to a descriptor pkh(key)

P2TR

P2TR template. Expands to a descriptor tr(key)

P2Wpkh

P2WPKH template. Expands to a descriptor wpkh(key)

P2Wpkh_P2Sh

P2WPKH-P2SH template. Expands to a descriptor sh(wpkh(key))

Traits

DescriptorTemplate

Trait for descriptor templates that can be built into a full descriptor

Type Aliases

DescriptorTemplateOut

Type alias for the return type of DescriptorTemplate, descriptor! and others