bdk_wallet::descriptor
Trait IntoWalletDescriptor

pub trait IntoWalletDescriptor {
    // Required method
    fn into_wallet_descriptor(
        self,
        secp: &Secp256k1<All>,
        network: Network,
    ) -> Result<(ExtendedDescriptor, KeyMap), DescriptorError>;
}

Trait for types which can be converted into an ExtendedDescriptor and a KeyMap usable by a wallet in a specific [Network]

Required Methods

fn into_wallet_descriptor( self, secp: &Secp256k1<All>, network: Network, ) -> Result<(ExtendedDescriptor, KeyMap), DescriptorError>

Convert to wallet descriptor