bdk_wallet::keys
Enum DescriptorPublicKey

pub enum DescriptorPublicKey {
    Single(SinglePub),
    XPub(DescriptorXKey<Xpub>),
    MultiXPub(DescriptorMultiXKey<Xpub>),
}

The descriptor pubkey, either a single pubkey or an xpub.

Variants

Single(SinglePub)

Single public key.

XPub(DescriptorXKey<Xpub>)

Extended public key (xpub).

MultiXPub(DescriptorMultiXKey<Xpub>)

Multiple extended public keys.

Implementations

impl DescriptorPublicKey

pub fn master_fingerprint(&self) -> Fingerprint

The fingerprint of the master key associated with this key, 0x00000000 if none.

pub fn full_derivation_path(&self) -> Option<DerivationPath>

Full path, from the master key

For wildcard keys this will return the path up to the wildcard, so you can get full paths by appending one additional derivation step, according to the wildcard type (hardened or normal).

For multipath extended keys, this returns None.

pub fn full_derivation_paths(&self) -> Vec<DerivationPath>

Returns a vector containing the full derivation paths from the master key. The vector will contain just one element for single keys, and multiple elements for multipath extended keys.

For wildcard keys this will return the path up to the wildcard, so you can get full paths by appending one additional derivation step, according to the wildcard type (hardened or normal).

pub fn is_deriveable(&self) -> bool

๐Deprecated: use has_wildcard instead

Whether or not the key has a wildcard

pub fn has_wildcard(&self) -> bool

Whether or not the key has a wildcard

pub fn derive( self, index: u32, ) -> Result<DefiniteDescriptorKey, ConversionError>

๐Deprecated: use at_derivation_index instead

Deprecated name for Self::at_derivation_index.

pub fn at_derivation_index( self, index: u32, ) -> Result<DefiniteDescriptorKey, ConversionError>

Replaces any wildcard (i.e. /*) in the key with a particular derivation index, turning it into a definite key (i.e. one where all the derivation paths are set).

Returns

If this key is not an xpub, returns self.

If this key is an xpub but does not have a wildcard, returns self.

Otherwise, returns the xpub at derivation index (removing the wildcard).

Errors

If index is hardened.

pub fn is_multipath(&self) -> bool

Whether or not this key has multiple derivation paths.

pub fn into_single_keys(self) -> Vec<DescriptorPublicKey>

Get as many keys as derivation paths in this key.

For raw public key and single-path extended keys it will return the key itself. For multipath extended keys it will return a single-path extended key per derivation path.