bdk_wallet::keys
Enum ExtendedKey

pub enum ExtendedKey<Ctx: ScriptContext = Legacy> {
    Private((Xpriv, PhantomData<Ctx>)),
    Public((Xpub, PhantomData<Ctx>)),
}

Enum for extended keys that can be either xprv or xpub

An instance of ExtendedKey can be constructed from an Xpriv or an Xpub by using the From trait.

Defaults to the Legacy context.

Variants

Private((Xpriv, PhantomData<Ctx>))

A private extended key, aka an xprv

Public((Xpub, PhantomData<Ctx>))

A public extended key, aka an xpub

Implementations

impl<Ctx: ScriptContext> ExtendedKey<Ctx>

pub fn has_secret(&self) -> bool

Return whether or not the key contains the private data

pub fn into_xprv(self, network: Network) -> Option<Xpriv>

Transform the ExtendedKey into an Xpriv for the given [Network], if the key contains the private data

pub fn into_xpub<C: Signing>( self, network: Network, secp: &Secp256k1<C>, ) -> Xpub

Transform the ExtendedKey into an Xpub for the given [Network]