Trait bdk_wallet::keys::GeneratableDefaultOptions

pub trait GeneratableDefaultOptions<Ctx>: GeneratableKey<Ctx>
where
    Ctx: ScriptContext,
    <Self as GeneratableKey<Ctx>>::Options: Default,
{
    // Provided methods
    fn generate_with_entropy_default(
        entropy: Self::Entropy
    ) -> Result<GeneratedKey<Self, Ctx>, Self::Error> { ... }
    fn generate_default() -> Result<GeneratedKey<Self, Ctx>, Self::Error> { ... }
    fn generate_default_with_aux_rand(
        rng: &mut (impl CryptoRng + RngCore)
    ) -> Result<GeneratedKey<Self, Ctx>, Self::Error> { ... }
}

Trait that allows generating a key with the default options

This trait is automatically implemented if the GeneratableKey::Options implements Default.

Provided Methods

fn generate_with_entropy_default( entropy: Self::Entropy ) -> Result<GeneratedKey<Self, Ctx>, Self::Error>

Generate a key with the default options and a given entropy

fn generate_default() -> Result<GeneratedKey<Self, Ctx>, Self::Error>

Generate a key with the default options and a random entropy

Uses the thread-local random number generator.

fn generate_default_with_aux_rand( rng: &mut (impl CryptoRng + RngCore) ) -> Result<GeneratedKey<Self, Ctx>, Self::Error>

Generate a key with the default options and a random entropy

Uses a provided random number generator (rng).

Object Safety

This trait is not object safe.