Trait bdk_wallet::signer::InputSigner

pub trait InputSigner: SignerCommon {
    // Required method
    fn sign_input(
        &self,
        psbt: &mut Psbt,
        input_index: usize,
        sign_options: &SignOptions,
        secp: &Secp256k1<All>
    ) -> Result<(), SignerError>;
}

PSBT Input signer

This trait can be implemented to provide custom signers to the wallet. If the signer supports signing individual inputs, this trait should be implemented and BDK will provide automatically an implementation for TransactionSigner.

Required Methods

fn sign_input( &self, psbt: &mut Psbt, input_index: usize, sign_options: &SignOptions, secp: &Secp256k1<All> ) -> Result<(), SignerError>

Sign a single psbt input