bdk_wallet
Struct CreateParams

pub struct CreateParams { /* private fields */ }

Parameters for Wallet::create or PersistedWallet::create.

Implementations

impl CreateParams

pub fn new_single<D: IntoWalletDescriptor + Send + 'static>( descriptor: D, ) -> Self

Construct parameters with provided descriptor.

Default values:

change_descriptor = None

network = [Network::Bitcoin]

genesis_hash = None

lookahead = [DEFAULT_LOOKAHEAD]

Use this method only when building a wallet with a single descriptor. See also Wallet::create_single.

pub fn new<D: IntoWalletDescriptor + Send + 'static>( descriptor: D, change_descriptor: D, ) -> Self

Construct parameters with provided descriptor and change_descriptor.

Default values:

network = [Network::Bitcoin]

genesis_hash = None

lookahead = [DEFAULT_LOOKAHEAD]

pub fn keymap(self, keychain: KeychainKind, keymap: KeyMap) -> Self

Extend the given keychain’s keymap.

pub fn network(self, network: Network) -> Self

Set network.

pub fn genesis_hash(self, genesis_hash: BlockHash) -> Self

Use a custom genesis_hash.

pub fn lookahead(self, lookahead: u32) -> Self

Use a custom lookahead value.

The lookahead defines a number of script pubkeys to derive over and above the last revealed index. Without a lookahead the indexer will miss outputs you own when processing transactions whose output script pubkeys lie beyond the last revealed index. In most cases the default value [DEFAULT_LOOKAHEAD] is sufficient.

pub fn create_wallet<P>( self, persister: &mut P, ) -> Result<PersistedWallet<P>, CreateWithPersistError<P::Error>>
where P: WalletPersister,

Create PersistedWallet with the given WalletPersister.

pub async fn create_wallet_async<P>( self, persister: &mut P, ) -> Result<PersistedWallet<P>, CreateWithPersistError<P::Error>>
where P: AsyncWalletPersister,

Create PersistedWallet with the given AsyncWalletPersister.

pub fn create_wallet_no_persist(self) -> Result<Wallet, DescriptorError>

Create Wallet without persistence.