Struct bdk_wallet::PersistedWallet

pub struct PersistedWallet<P> { /* private fields */ }

Represents a persisted wallet which persists into type P.

This is a light wrapper around Wallet that enforces some level of safety-checking when used with a WalletPersister or AsyncWalletPersister implementation. Safety checks assume that WalletPersister and/or AsyncWalletPersister are implemented correctly.

Checks include:

Ensure the persister is initialized before data is persisted.

Ensure there were no previously persisted wallet data before creating a fresh wallet and persisting it.

Only clear the staged changes of Wallet after persisting succeeds.

Ensure the wallet is persisted to the same P type as when created/loaded. Note that this is not completely fool-proof as you can have multiple instances of the same P type that are connected to different databases.

Implementations

impl<P: WalletPersister> PersistedWallet<P>

Methods when P is a WalletPersister.

pub fn create( persister: &mut P, params: CreateParams ) -> Result<Self, CreateWithPersistError<P::Error>>

Create a new PersistedWallet with the given persister and params.

pub fn load( persister: &mut P, params: LoadParams ) -> Result<Option<Self>, LoadWithPersistError<P::Error>>

Load a previously PersistedWallet from the given persister and params.

pub fn persist(&mut self, persister: &mut P) -> Result<bool, P::Error>

Persist staged changes of wallet into persister.

Returns whether any new changes were persisted.

If the persister errors, the staged changes will not be cleared.

impl<P: AsyncWalletPersister> PersistedWallet<P>

Methods when P is an AsyncWalletPersister.

pub async fn create_async( persister: &mut P, params: CreateParams ) -> Result<Self, CreateWithPersistError<P::Error>>

Create a new PersistedWallet with the given async persister and params.

pub async fn load_async( persister: &mut P, params: LoadParams ) -> Result<Option<Self>, LoadWithPersistError<P::Error>>

Load a previously PersistedWallet from the given async persister and params.

pub async fn persist_async<'a>( &'a mut self, persister: &mut P ) -> Result<bool, P::Error>

Persist staged changes of wallet into an async persister.

Returns whether any new changes were persisted.

If the persister errors, the staged changes will not be cleared.