bdk_wallet
Trait IsDust

pub trait IsDust {
    // Required method
    fn is_dust(&self, script: &Script) -> bool;
}

Trait to check if a value is below the dust limit. We are performing dust value calculation for a given script public key using rust-bitcoin to keep it compatible with network dust rate

Required Methods

fn is_dust(&self, script: &Script) -> bool

Check whether or not a value is below dust limit